--- frontend/app/lib/queries.ts
@@ -1,3198 +1,0 @@
-import {
-  type InfiniteData,
-  type QueryClient,
-  infiniteQueryOptions,
-  keepPreviousData,
-  queryOptions,
-  type skipToken
-} from "@tanstack/react-query";
-import { href, redirect } from "react-router";
-import { preprocess, z } from "zod";
-import { zfd } from "zod-form-data";
-import type { ApiResponse, RequestParams } from "~/api/client";
-import { apiClient } from "~/api/client";
-import type {
-  TemplateDetailsApiResponse,
-  TemplateSearchAPIResponse
-} from "~/api/types";
-import {
-  DEFAULT_LOGS_PER_PAGE,
-  DEFAULT_QUERY_REFETCH_INTERVAL,
-  DEPLOYMENT_STATUSES,
-  LOGS_QUERY_REFETCH_INTERVAL,
-  METRICS_TIME_RANGES,
-  TEMPLATE_API_HOST,
-  WORKSPACE_ROLE_MAPPING
-} from "~/lib/constants";
-import type { Writeable } from "~/lib/types";
-import { durationToMs, hasMinRole, notFound } from "~/lib/utils";
-
-export const userQueries = {
-  authedUser: queryOptions({
-    queryKey: ["AUTHED_USER"] as const,
-    queryFn: async ({ signal }) => {
-      const { data } = await apiClient.GET("/api/auth/me/", { signal });
-      return data ?? null;
-    },
-    refetchInterval: (query) => {
-      if (query.state.data) {
-        return durationToMs(30, "minutes");
-      }
-      return false;
-    }
-  }),
-
-  memberships: queryOptions({
-    queryKey: ["WORKSPACE_MEMBERSHIP", "LIST"] as const,
-    queryFn: async ({ signal }) => {
-      const { data } = await apiClient.GET("/api/workspaces/list/", { signal });
-      return data ?? null;
-    },
-    refetchInterval: (query) => {
-      if (query.state.data) {
-        return durationToMs(30, "minutes");
-      }
-      return false;
-    }
-  }),
-
-  checkUserExistence: queryOptions({
-    queryKey: ["CHECK_USER_EXISTENCE"] as const,
-    queryFn: async ({ signal }) => {
-      const result = await apiClient.GET("/api/auth/check-user-existence/", {
-        signal
-      });
-
-      // if rate limited, throw error
-      if (result.response.status === 429) {
-        const fullErrorMessage = result.error?.errors
-          .map((err) => err.detail)
-          .join(" ");
-
-        throw new Error(fullErrorMessage);
-      }
-
-      return result;
-    }
-  })
-};
-
-/**
- * Fetches the authed user and redirects to `/login` if there isn't one.
- * Centralizes the "what if there isn't one" case instead of leaving it to
- * each `clientLoader` to null-check and redirect on its own.
- */
-export async function ensureAuthedUser(queryClient: QueryClient) {
-  const user = await queryClient.ensureQueryData(userQueries.authedUser);
-  if (!user) {
-    throw redirect(href("/login"));
-  }
-  return user;
-}
-
-/**
- * Fetches the authed user and throws a 404 if they don't have at least
- * `roleName` in the current workspace. Centralizes the auth + role check
- * instead of leaving it to each `clientLoader` to do both on its own.
- */
-export async function ensureMinRole(
-  queryClient: QueryClient,
-  roleName: Parameters<typeof hasMinRole>[1]
-) {
-  const user = await ensureAuthedUser(queryClient);
-  if (!hasMinRole(user, roleName)) {
-    throw notFound(
-      import.meta.env.DEV
-        ? "You do have permission to view this page"
-        : "Not found"
-    );
-  }
-  return user;
-}
-
-export const dockerHubQueries = {
-  images: (query: string) =>
-    queryOptions({
-      queryKey: ["DOCKER_HUB_IMAGES", query] as const,
-      queryFn: ({ signal }) => {
-        return apiClient.GET("/api/docker/image-search/", {
-          params: {
-            query: {
-              q: query.trim()
-            }
-          },
-          signal
-        });
-      },
-      enabled: query.trim().length > 0
-    })
-};
-
-export const projectSearchSchema = zfd.formData({
-  slug: z.string().optional().catch(undefined),
-  sort_by: zfd
-    .repeatable(z.array(z.enum(["slug", "-updated_at"])))
-    .optional()
-    .catch(undefined)
-});
-
-export type ProjectSearch = z.infer<typeof projectSearchSchema>;
-
-/**
- * Base query key for all resources scoped to a workspace.
- * The current workspace is server-side session state, so the react-query
- * cache must be partitioned by `workspaceId` to avoid leaking data across
- * workspaces.
- */
-export const workspaceKey = (workspaceId: string) =>
-  ["WORKSPACE", workspaceId] as const;
-
-export const projectQueries = {
-  list: ({
-    workspaceId,
-    filters = {},
-    refetchInterval = DEFAULT_QUERY_REFETCH_INTERVAL
-  }: {
-    workspaceId: string;
-    filters?: ProjectSearch;
-    refetchInterval?: number;
-  }) =>
-    queryOptions({
-      queryKey: [
-        ...workspaceKey(workspaceId),
-        "PROJECT_LIST",
-        filters
-      ] as const,
-      queryFn: async ({ signal }) => {
-        const { data } = await apiClient.GET("/api/projects/", {
-          params: {
-            query: {
-              ...filters
-            }
-          },
-          signal
-        });
-        if (!data) {
-          throw notFound(`Not found`);
-        }
-        return data;
-      },
-      placeholderData: keepPreviousData,
-      refetchInterval: (query) => {
-        if (query.state.data) {
-          return refetchInterval;
-        }
-        return false;
-      }
-    }),
-  single: (workspaceId: string, slug: string) =>
-    queryOptions({
-      queryKey: [...workspaceKey(workspaceId), "PROJECT_SINGLE", slug] as const,
-      queryFn: async ({ signal }) => {
-        const { data } = await apiClient.GET("/api/projects/{slug}/", {
-          params: {
-            path: {
-              slug
-            }
-          },
-          signal
-        });
-        if (!data) {
-          throw notFound(
-            `The project \`${slug}\` does not exist on this workspace`
-          );
-        }
-        return data;
-      },
-      placeholderData: keepPreviousData
-    })
-};
-
-export const environmentQueries = {
-  single: (workspaceId: string, project_slug: string, env_slug: string) =>
-    queryOptions({
-      queryKey: [
-        ...projectQueries.single(workspaceId, project_slug).queryKey,
-        env_slug
-      ] as const,
-      queryFn: async ({ signal }) => {
-        const { data } = await apiClient.GET(
-          "/api/projects/{slug}/environment-details/{env_slug}/",
-          {
-            params: {
-              path: {
-                slug: project_slug,
-                env_slug
-              }
-            },
-            signal
-          }
-        );
-        if (!data) {
-          throw notFound();
-        }
-        return data;
-      },
-      refetchInterval: (query) => {
-        if (query.state.data) {
-          return DEFAULT_QUERY_REFETCH_INTERVAL;
-        }
-        return false;
-      }
-    }),
-
-  serviceList: (
-    workspaceId: string,
-    project_slug: string,
-    env_slug: string,
-    filters: ProjectServiceListSearch = {}
-  ) =>
-    queryOptions({
-      queryKey: [
-        ...environmentQueries.single(workspaceId, project_slug, env_slug)
-          .queryKey,
-        "SERVICE_LIST",
-        filters
-      ] as const,
-      queryFn: async ({ signal }) => {
-        const { data } = await apiClient.GET(
-          "/api/projects/{slug}/{env_slug}/service-list/",
-          {
-            params: {
-              query: {
-                ...filters
-              },
-              path: {
-                slug: project_slug,
-                env_slug
-              }
-            },
-            signal
-          }
-        );
-
-        if (!data) {
-          throw notFound();
-        }
-        return data;
-      },
-      refetchInterval: (query) => {
-        if (query.state.data) {
-          return DEFAULT_QUERY_REFETCH_INTERVAL;
-        }
-        return false;
-      }
-    }),
-
-  composeStackList: (
-    workspaceId: string,
-    project_slug: string,
-    env_slug: string,
-    filters: ProjectSearch = {}
-  ) =>
-    queryOptions({
-      queryKey: [
-        ...environmentQueries.single(workspaceId, project_slug, env_slug)
-          .queryKey,
-        "COMPOSE_STACK_LIST",
-        filters
-      ] as const,
-      queryFn: async ({ signal }) => {
-        const { data } = await apiClient.GET(
-          "/api/compose/stacks/{project_slug}/{env_slug}/",
-          {
-            params: {
-              query: {
-                ...filters
-              },
-              path: {
-                project_slug,
-                env_slug
-              }
-            },
-            signal
-          }
-        );
-
-        if (!data) {
-          throw notFound();
-        }
-        return data;
-      },
-      refetchInterval: (query) => {
-        if (query.state.data) {
-          return DEFAULT_QUERY_REFETCH_INTERVAL;
-        }
-        return false;
-      }
-    }),
-
-  pendingReview: (
-    workspaceId: string,
-    project_slug: string,
-    env_slug: string
-  ) =>
-    queryOptions({
-      queryKey: [
-        ...projectQueries.single(workspaceId, project_slug).queryKey,
-        env_slug,
-        "PENDING_REVIEW"
-      ] as const,
-      queryFn: async ({ signal }) => {
-        const { data } = await apiClient.GET(
-          "/api/projects/{slug}/environment-details/{env_slug}/review-preview-deployment/",
-          {
-            params: {
-              path: {
-                slug: project_slug,
-                env_slug
-              }
-            },
-            signal
-          }
-        );
-        if (!data) {
-          throw notFound(
-            `No pending environment to review exists at \`${project_slug}/${env_slug}\` `
-          );
-        }
-        return data;
-      },
-      refetchInterval: (query) => {
-        if (query.state.data) {
-          return DEFAULT_QUERY_REFETCH_INTERVAL;
-        }
-        return false;
-      }
-    })
-};
-
-export const projectServiceListSearchSchema = z.object({
-  query: z.string().optional().catch("")
-});
-export type ProjectServiceListSearch = z.infer<
-  typeof projectServiceListSearchSchema
->;
-
-export const serviceDeploymentListFilters = zfd.formData({
-  page: zfd.numeric().optional().catch(1).optional(),
-  per_page: zfd.numeric().optional().catch(10).optional(),
-  status: zfd.repeatable(
-    z
-      .array(z.enum(DEPLOYMENT_STATUSES))
-      .optional()
-      .catch(DEPLOYMENT_STATUSES as Writeable<typeof DEPLOYMENT_STATUSES>)
-  ),
-  queued_at_before: z.coerce.date().optional().catch(undefined),
-  queued_at_after: z.coerce.date().optional().catch(undefined)
-});
-
-export type ServiceDeploymentListFilters = z.infer<
-  typeof serviceDeploymentListFilters
->;
-
-export const metrisSearch = z.object({
-  time_range: z
-    .enum(METRICS_TIME_RANGES)
-    .optional()
-    .default("LAST_HOUR")
-    .catch("LAST_HOUR")
-});
-
-export type MetricsFilters = z.TypeOf<typeof metrisSearch>;
-
-/************************************
- *       Compose Stack Queries      *
- ************************************/
-
-export const stackDeploymentListFilters = zfd.formData({
-  page: zfd.numeric().optional().catch(1).optional(),
-  per_page: zfd.numeric().optional().catch(10).optional(),
-  status: zfd.repeatable(
-    z
-      .array(z.enum(["QUEUED", "DEPLOYING", "FINISHED", "FAILED", "CANCELLED"]))
-      .optional()
-      .catch(undefined)
-  ),
-  queued_at_before: z.coerce.date().optional().catch(undefined),
-  queued_at_after: z.coerce.date().optional().catch(undefined)
-});
-
-export type StackDeploymentListFilters = z.infer<
-  typeof stackDeploymentListFilters
->;
-
-export const stackMetrisSearch = z.object({
-  time_range: z
-    .enum(METRICS_TIME_RANGES)
-    .optional()
-    .default("LAST_HOUR")
-    .catch("LAST_HOUR"),
-  service_names: zfd.repeatable(z.array(z.string()))
-});
-
-export type StackMetricsFilters = z.TypeOf<typeof stackMetrisSearch>;
-
-export const LOG_LEVELS = ["INFO", "ERROR"] as const;
-
-export const stackRuntimeLogSearchSchema = zfd.formData({
-  level: zfd.repeatable(
-    z
-      .array(z.enum(LOG_LEVELS))
-      .optional()
-      .catch(LOG_LEVELS as Writeable<typeof LOG_LEVELS>)
-  ),
-  time_before: z.coerce.date().optional().catch(undefined),
-  time_after: z.coerce.date().optional().catch(undefined),
-  content: z.string().optional(),
-  query: z.string().optional(),
-  container_id: z.string().optional(),
-  isMaximized: preprocess(
-    (arg) => arg === "true",
-    z.coerce.boolean().optional().catch(false)
-  ),
-  context: z.coerce.number().optional().catch(undefined),
-  context_lines: z.coerce.number().min(5).optional().catch(undefined)
-});
-
-export type ComposeStackRuntimeLogFilters = z.infer<
-  typeof stackRuntimeLogSearchSchema
->;
-
-export const composeStackQueries = {
-  single: ({
-    workspaceId,
-    project_slug,
-    stack_slug,
-    env_slug
-  }: {
-    workspaceId: string;
-    project_slug: string;
-    env_slug: string;
-    stack_slug: string;
-  }) =>
-    queryOptions({
-      queryKey: [
-        ...projectQueries.single(workspaceId, project_slug).queryKey,
-        env_slug,
-        "COMPOSE_STACK_DETAILS",
-        stack_slug
-      ] as const,
-      queryFn: async ({ signal }) => {
-        const { data } = await apiClient.GET(
-          "/api/compose/stacks/{project_slug}/{env_slug}/{slug}/",
-          {
-            params: {
-              path: {
-                project_slug,
-                slug: stack_slug,
-                env_slug
-              }
-            },
-            signal
-          }
-        );
-
-        if (!data) {
-          throw notFound(
-            `The compose stack \`${stack_slug}\` doesn't exist in this project.`
-          );
-        }
-        return data;
-      },
-      refetchInterval: (query) => {
-        if (query.state.data) {
-          return DEFAULT_QUERY_REFETCH_INTERVAL;
-        }
-        return false;
-      },
-      refetchIntervalInBackground: true
-    }),
-  singleDeployment: ({
-    workspaceId,
-    project_slug,
-    stack_slug,
-    env_slug,
-    deployment_hash
-  }: {
-    workspaceId: string;
-    project_slug: string;
-    env_slug: string;
-    stack_slug: string;
-    deployment_hash: string;
-  }) =>
-    queryOptions({
-      queryKey: [
-        ...composeStackQueries.deploymentList({
-          workspaceId,
-          project_slug,
-          env_slug,
-          stack_slug
-        }).queryKey,
-        "DETAILS",
-        deployment_hash
-      ] as const,
-      queryFn: async ({ signal }) => {
-        const { data } = await apiClient.GET(
-          "/api/compose/stacks/{project_slug}/{env_slug}/{slug}/deployments/{hash}/",
-          {
-            params: {
-              path: {
-                project_slug,
-                slug: stack_slug,
-                env_slug,
-                hash: deployment_hash
-              }
-            },
-            signal
-          }
-        );
-
-        if (!data) {
-          throw notFound(
-            `The Deployment \`${stack_slug}\` doesn't exist in this stack.`
-          );
-        }
-        return data;
-      },
-      refetchInterval: (query) => {
-        if (query.state.data) {
-          return DEFAULT_QUERY_REFETCH_INTERVAL;
-        }
-        return false;
-      },
-      refetchIntervalInBackground: true
-    }),
-  deploymentLogs: ({
-    workspaceId,
-    project_slug,
-    stack_slug,
-    env_slug,
-    deployment_hash,
-    queryClient,
-    autoRefetchEnabled
-  }: {
-    workspaceId: string;
-    project_slug: string;
-    env_slug: string;
-    stack_slug: string;
-    deployment_hash: string;
-    queryClient: QueryClient;
-    autoRefetchEnabled?: boolean;
-  }) =>
-    infiniteQueryOptions({
-      queryKey: [
-        ...composeStackQueries.singleDeployment({
-          workspaceId,
-          project_slug,
-          stack_slug,
-          env_slug,
-          deployment_hash
-        }).queryKey,
-        "BUILD_LOGS"
-      ],
-      queryFn: async ({ pageParam, signal, queryKey }) => {
-        const allData = queryClient.getQueryData(queryKey) as InfiniteData<
-          DeploymentLogQueryData,
-          string | null
-        >;
-        const existingData = allData?.pages.find(
-          (_, index) => allData?.pageParams[index] === pageParam
-        );
-
-        /**
-         * We reuse the data in the query as we are sure this page is immutable,
-         * And we don't want to refetch the same logs that we have already fetched.
-         *
-         * However if we have the data in the cache and next is `null`,
-         * it means that that page is the last page with the most recent data
-         * and the next time we fetch it, there might be more data available.
-         * Inspired by: https://github.com/TanStack/query/discussions/5921
-         */
-        if (existingData?.next) {
-          return existingData;
-        }
-
-        /**
-         * when we issue a refetch, for all pages we fetched via `fetchPreviousPage` starting from the second page,
-         * tanstack query will use the `next` page pointer of the previous to refetch them,
-         * so we check if we already have it.
-         * In the docs, it's so that the data the pointers aren't stale, but we don't have that issue
-         * since the log data is immutable.
-         * ref: https://tanstack.com/query/latest/docs/framework/react/guides/infinite-queries#what-happens-when-an-infinite-query-needs-to-be-refetched
-         */
-        const existingDataIndex = allData?.pages.findIndex(
-          (_, index) => allData?.pages[index].next === pageParam
-        );
-        if (!existingData && existingDataIndex > -1) {
-          const nextPage = allData.pages[existingDataIndex + 1];
-          if (nextPage) {
-            return nextPage;
-          }
-        }
-
-        // the actual request
-        const { data } = await apiClient.GET(
-          "/api/compose/stacks/{project_slug}/{env_slug}/{slug}/deployments/{hash}/build-logs/",
-          {
-            params: {
-              path: {
-                project_slug,
-                slug: stack_slug,
-                env_slug,
-                hash: deployment_hash
-              },
-              query: {
-                per_page: DEFAULT_LOGS_PER_PAGE,
-                cursor: pageParam ?? existingData?.cursor ?? undefined
-              }
-            },
-            signal
-          }
-        );
-
-        let apiData: DeploymentLogQueryData = {
-          next: null,
-          previous: null,
-          results: [],
-          cursor: null
-        };
-
-        if (data) {
-          // we reverse the results and reverse the page pointers (next/previous) because
-          // the data from the API is in reverse order of traversal and timestamp.
-          // Reversing them allows us to reorder the data in the ascending order as it is shown in the UI
-          apiData = {
-            results: data.results.toReversed(),
-            next: data?.previous ?? null,
-            previous: data?.next ?? null,
-            cursor: existingData?.cursor
-          };
-        }
-
-        // get cursor for initial page as its pageParam is `null`
-        // we want to do that because we don't to always fetch the latest data for the initial page
-        // instead what we want is to fetch from the time it starts
-        if (
-          pageParam === null &&
-          !apiData.cursor &&
-          !apiData.next &&
-          apiData.results.length > 0
-        ) {
-          const oldestLog = apiData.results[0];
-          const cursor = { sort: [oldestLog.timestamp], order: "asc" };
-          apiData.cursor = btoa(JSON.stringify(cursor));
-        }
-
-        return apiData;
-      },
-      getNextPageParam: ({ next }) => next,
-      getPreviousPageParam: ({ previous }) => previous,
-      initialPageParam: null as string | null,
-      refetchInterval: (query) => {
-        if (!query.state.data || !autoRefetchEnabled) {
-          return false;
-        }
-        return LOGS_QUERY_REFETCH_INTERVAL;
-      },
-      placeholderData: keepPreviousData,
-      staleTime: Number.POSITIVE_INFINITY
-    }),
-  deploymentList: ({
-    workspaceId,
-    project_slug,
-    stack_slug,
-    env_slug,
-    filters = {}
-  }: {
-    workspaceId: string;
-    project_slug: string;
-    env_slug: string;
-    stack_slug: string;
-    filters?: StackDeploymentListFilters;
-  }) =>
-    queryOptions({
-      queryKey: [
-        ...composeStackQueries.single({
-          workspaceId,
-          project_slug,
-          env_slug,
-          stack_slug
-        }).queryKey,
-        "DEPLOYMENT_LIST",
-        filters
-      ] as const,
-      queryFn: async ({ signal }) => {
-        const { data } = await apiClient.GET(
-          "/api/compose/stacks/{project_slug}/{env_slug}/{slug}/deployments/",
-          {
-            params: {
-              path: {
-                project_slug,
-                slug: stack_slug,
-                env_slug
-              },
-              query: {
-                ...filters,
-                queued_at_after: filters.queued_at_after?.toISOString(),
-                queued_at_before: filters.queued_at_before?.toISOString()
-              }
-            },
-            signal
-          }
-        );
-
-        if (!data) {
-          throw notFound(
-            `The compose stack \`${stack_slug}\` doesn't exist in this project.`
-          );
-        }
-        return data;
-      },
-      refetchInterval: (query) => {
-        if (query.state.data) {
-          return DEFAULT_QUERY_REFETCH_INTERVAL;
-        }
-        return false;
-      }
-    }),
-  httpLogs: ({
-    workspaceId,
-    project_slug,
-    stack_slug,
-    env_slug,
-    stack_id,
-    queryClient,
-    stack_service_name,
-    autoRefetchEnabled = true,
-    filters = {}
-  }: {
-    workspaceId: string;
-    stack_id: string;
-    project_slug: string;
-    stack_slug: string;
-    env_slug: string;
-    filters?: Omit<HTTPLogFilters, "isMaximized">;
-    queryClient: QueryClient;
-    autoRefetchEnabled?: boolean;
-    stack_service_name?: string[];
-  }) =>
-    infiniteQueryOptions({
-      queryKey: [
-        ...composeStackQueries.single({
-          workspaceId,
-          project_slug,
-          stack_slug,
-          env_slug
-        }).queryKey,
-        "HTTP_LOGS",
-        stack_id,
-        stack_service_name,
-        filters
-      ] as const,
-      queryFn: async ({ pageParam, signal, queryKey }) => {
-        const allData = queryClient.getQueryData(queryKey) as InfiniteData<
-          HttpLogQueryData,
-          string | null
-        >;
-        const existingData = allData?.pages.find(
-          (_, index) => allData?.pageParams[index] === pageParam
-        );
-
-        /**
-         * We reuse the data in the query as we are sure this page is immutable,
-         * And we don't want to refetch the same logs that we have already fetched.
-         *
-         * However if we have the data in the cache and previous is `null`,
-         * it means that that page is the last and the next time we fetch it,
-         * it might have more data.
-         * Inspired by: https://github.com/TanStack/query/discussions/5921
-         */
-        if (existingData?.previous) {
-          return existingData;
-        }
-
-        let cursor = pageParam ?? undefined;
-        if (existingData?.cursor) {
-          cursor = existingData.cursor;
-        }
-
-        const { data } = await apiClient.GET("/api/http-logs/", {
-          params: {
-            query: {
-              ...filters,
-              cursor,
-              stack_id,
-              stack_service_name:
-                (stack_service_name ?? []).length === 0
-                  ? undefined
-                  : stack_service_name,
-              per_page: DEFAULT_LOGS_PER_PAGE,
-              time_before: filters.time_before?.toISOString(),
-              time_after: filters.time_after?.toISOString()
-            }
-          },
-          signal
-        });
-
-        let apiData: HttpLogQueryData = {
-          next: null,
-          previous: null,
-          results: [],
-          cursor: null
-        };
-
-        if (data) {
-          const next = data.next
-            ? new URL(data.next).searchParams.get("cursor")
-            : null;
-          const previous = data.previous
-            ? new URL(data.previous).searchParams.get("cursor")
-            : null;
-          apiData = {
-            results: data.results,
-            next,
-            previous,
-            cursor: existingData?.cursor
-          };
-        }
-
-        // get cursor for initial page as its pageParam is `null`
-        // we want to do so that we don't to always fetch the latest data for the initial page
-        // instead what we want is to fetch from the data it starts
-        if (pageParam === null && apiData.next !== null && !apiData.cursor) {
-          const { data: nextPage } = await apiClient.GET("/api/http-logs/", {
-            params: {
-              query: {
-                ...filters,
-                stack_id,
-                stack_service_name:
-                  (stack_service_name ?? []).length === 0
-                    ? undefined
-                    : stack_service_name,
-                per_page: DEFAULT_LOGS_PER_PAGE,
-                cursor: apiData.next,
-                time_before: filters.time_before?.toISOString(),
-                time_after: filters.time_after?.toISOString()
-              }
-            },
-            signal
-          });
-          if (nextPage?.previous) {
-            apiData.cursor = new URL(nextPage.previous).searchParams.get(
-              "cursor"
-            );
-          }
-        }
-
-        return apiData;
-      },
-      refetchInterval: (query) => {
-        if (!query.state.data || !autoRefetchEnabled) {
-          return false;
-        }
-        return DEFAULT_QUERY_REFETCH_INTERVAL;
-      },
-      getNextPageParam: ({ next }) => next,
-      getPreviousPageParam: ({ previous }) => previous,
-      initialPageParam: null as string | null,
-      placeholderData: keepPreviousData,
-      staleTime: Number.POSITIVE_INFINITY
-    }),
-  singleHttpLog: ({
-    workspaceId,
-    project_slug,
-    stack_slug,
-    env_slug,
-    request_uuid
-  }: {
-    workspaceId: string;
-    project_slug: string;
-    stack_slug: string;
-    env_slug: string;
-    request_uuid: string;
-  }) =>
-    queryOptions({
-      queryKey: [
-        ...composeStackQueries.single({
-          workspaceId,
-          project_slug,
-          stack_slug,
-          env_slug
-        }).queryKey,
-        "HTTP_LOGS",
-        request_uuid
-      ] as const,
-      queryFn: async ({ signal }) => {
-        const { data } = await apiClient.GET("/api/http-logs/{request_uuid}/", {
-          params: {
-            path: {
-              request_uuid
-            }
-          },
-          signal
-        });
-        return data;
-      }
-    }),
-  filterHttpLogFields: ({
-    workspaceId,
-    project_slug,
-    env_slug,
-    stack_id,
-    stack_slug,
-    field,
-    value
-  }: {
-    workspaceId: string;
-    project_slug: string;
-    stack_id: string;
-    stack_slug: string;
-    env_slug: string;
-    field: RequestParams<"get", "/api/http-logs/fields/">["field"];
-    value: string;
-  }) =>
-    queryOptions({
-      queryKey: [
-        ...composeStackQueries.single({
-          workspaceId,
-          project_slug,
-          stack_slug,
-          env_slug
-        }).queryKey,
-        stack_id,
-        "HTTP_LOG_FIELDS",
-        field,
-        value
-      ],
-      queryFn: async ({ signal }) => {
-        const { data } = await apiClient.GET("/api/http-logs/fields/", {
-          signal,
-          params: {
-            query: {
-              field,
-              value,
-              stack_id
-            }
-          }
-        });
-        return data ?? [];
-      }
-    }),
-  metrics: ({
-    workspaceId,
-    project_slug,
-    stack_slug,
-    env_slug,
-    filters
-  }: {
-    workspaceId: string;
-    project_slug: string;
-    stack_slug: string;
-    env_slug: string;
-    filters?: StackMetricsFilters;
-  }) =>
-    queryOptions({
-      queryKey: [
-        ...composeStackQueries.single({
-          workspaceId,
-          project_slug,
-          stack_slug,
-          env_slug
-        }).queryKey,
-        "METRICS",
-        filters
-      ] as const,
-      queryFn: async ({ signal }) => {
-        const { data } = await apiClient.GET(
-          "/api/compose/stacks/{project_slug}/{env_slug}/{slug}/metrics/",
-          {
-            params: {
-              path: {
-                project_slug,
-                slug: stack_slug,
-                env_slug
-              },
-              query: {
-                ...filters
-              }
-            },
-            signal
-          }
-        );
-
-        if (!data) {
-          throw notFound();
-        }
-        return data;
-      },
-      refetchInterval: (query) => {
-        if (query.state.data) {
-          return DEFAULT_QUERY_REFETCH_INTERVAL;
-        }
-        return false;
-      }
-    }),
-  runtimeLogs: ({
-    workspaceId,
-    project_slug,
-    service_name,
-    stack_slug,
-    env_slug,
-    autoRefetchEnabled = true,
-    filters = {},
-    queryClient
-  }: {
-    workspaceId: string;
-    project_slug: string;
-    service_name: string;
-    stack_slug: string;
-    env_slug: string;
-    filters?: Omit<ComposeStackRuntimeLogFilters, "isMaximized">;
-    queryClient: QueryClient;
-    autoRefetchEnabled?: boolean;
-  }) =>
-    infiniteQueryOptions({
-      queryKey: [
-        ...composeStackQueries.single({
-          workspaceId,
-          project_slug,
-          stack_slug,
-          env_slug
-        }).queryKey,
-        "RUNTIME_LOGS",
-        service_name,
-        filters
-      ],
-      queryFn: async ({ pageParam, signal, queryKey }) => {
-        const allData = queryClient.getQueryData(queryKey) as InfiniteData<
-          DeploymentLogQueryData,
-          string | null
-        >;
-        const existingData = allData?.pages.find(
-          (_, index) => allData?.pageParams[index] === pageParam
-        );
-
-        /**
-         * We reuse the data in the query as we are sure this page is immutable,
-         * And we don't want to refetch the same logs that we have already fetched.
-         *
-         * However if we have the data in the cache and next is `null`,
-         * it means that that page is the last page with the most recent data
-         * and the next time we fetch it, there might be more data available.
-         * Inspired by: https://github.com/TanStack/query/discussions/5921
-         */
-        if (existingData?.next) {
-          return existingData;
-        }
-
-        /**
-         * when we issue a refetch, for all pages we fetched via `fetchPreviousPage` starting from the second page,
-         * tanstack query will use the `next` page pointer of the previous to refetch them,
-         * so we check if we already have it.
-         * In the docs, it's so that the data the pointers aren't stale, but we don't have that issue
-         * since the log data is immutable.
-         * ref: https://tanstack.com/query/latest/docs/framework/react/guides/infinite-queries#what-happens-when-an-infinite-query-needs-to-be-refetched
-         */
-        const existingDataIndex = allData?.pages.findIndex(
-          (_, index) => allData?.pages[index].next === pageParam
-        );
-        if (!existingData && existingDataIndex > -1) {
-          const nextPage = allData.pages[existingDataIndex + 1];
-          if (nextPage) {
-            return nextPage;
-          }
-        }
-
-        // the actual request
-        const { data } = await apiClient.GET(
-          "/api/compose/stacks/{project_slug}/{env_slug}/{slug}/runtime-logs/",
-          {
-            params: {
-              path: {
-                project_slug,
-                env_slug,
-                slug: stack_slug
-              },
-              query: {
-                stack_service_name: service_name,
-                per_page: DEFAULT_LOGS_PER_PAGE,
-                cursor: pageParam ?? existingData?.cursor ?? undefined,
-                ...filters,
-                time_before: filters.time_before?.toISOString(),
-                time_after: filters.time_after?.toISOString()
-              }
-            },
-            signal
-          }
-        );
-
-        let apiData: DeploymentLogQueryData = {
-          next: null,
-          previous: null,
-          results: [],
-          cursor: null
-        };
-
-        if (data) {
-          // we reverse the results and reverse the page pointers (next/previous) because
-          // the data from the API is in reverse order of traversal and timestamp.
-          // Reversing them allows us to reorder the data in the ascending order as it is shown in the UI
-          apiData = {
-            results: data.results.toReversed(),
-            next: data?.previous ?? null,
-            previous: data?.next ?? null,
-            cursor: existingData?.cursor
-          };
-        }
-
-        // get cursor for initial page as its pageParam is `null`
-        // we want to do that because we don't to always fetch the latest data for the initial page
-        // instead what we want is to fetch from the time it starts
-        if (
-          pageParam === null &&
-          !apiData.cursor &&
-          !apiData.next &&
-          apiData.results.length > 0
-        ) {
-          const oldestLog = apiData.results[0];
-          const cursor = { sort: [oldestLog.timestamp], order: "asc" };
-          apiData.cursor = btoa(JSON.stringify(cursor));
-        }
-
-        return apiData;
-      },
-      getNextPageParam: ({ next }) => next,
-      getPreviousPageParam: ({ previous }) => previous,
-      initialPageParam: null as string | null,
-      refetchInterval: (query) => {
-        if (!query.state.data || !autoRefetchEnabled) {
-          return false;
-        }
-        return LOGS_QUERY_REFETCH_INTERVAL;
-      },
-      placeholderData: keepPreviousData,
-      staleTime: Number.POSITIVE_INFINITY
-    }),
-  logWithContext: ({
-    workspaceId,
-    project_slug,
-    stack_slug,
-    env_slug,
-    service_name,
-    time,
-    context_lines = 20
-  }: {
-    workspaceId: string;
-    project_slug: string;
-    service_name: string;
-    stack_slug: string;
-    env_slug: string;
-    time: number;
-    context_lines?: number;
-  }) =>
-    queryOptions({
-      queryKey: [
-        ...composeStackQueries.single({
-          workspaceId,
-          project_slug,
-          stack_slug,
-          env_slug
-        }).queryKey,
-        "RUNTIME_LOGS",
-        "WITH_CONTEXT",
-        time,
-        context_lines
-      ],
-      queryFn: async ({ signal }) => {
-        const { data } = await apiClient.GET(
-          "/api/compose/stacks/{project_slug}/{env_slug}/{slug}/runtime-logs/with-context/{time}/",
-          {
-            params: {
-              path: {
-                project_slug,
-                slug: stack_slug,
-                env_slug,
-                time: time.toString()
-              },
-              query: {
-                lines: context_lines,
-                stack_service_name: service_name
-              }
-            },
-            signal
-          }
-        );
-
-        return data;
-      },
-      refetchInterval: (query) => {
-        if (!query.state.data) {
-          return false;
-        }
-        return LOGS_QUERY_REFETCH_INTERVAL;
-      },
-      placeholderData: keepPreviousData
-    })
-};
-
-export const serviceQueries = {
-  single: ({
-    workspaceId,
-    project_slug,
-    service_slug,
-    env_slug
-  }: {
-    workspaceId: string;
-    project_slug: string;
-    env_slug: string;
-    service_slug: string;
-  }) =>
-    queryOptions({
-      queryKey: [
-        ...projectQueries.single(workspaceId, project_slug).queryKey,
-        env_slug,
-        "SERVICE_DETAILS",
-        service_slug
-      ] as const,
-      queryFn: async ({ signal }) => {
-        const { data } = await apiClient.GET(
-          "/api/projects/{project_slug}/{env_slug}/service-details/{slug}/",
-          {
-            params: {
-              path: {
-                project_slug,
-                slug: service_slug,
-                env_slug
-              }
-            },
-            signal
-          }
-        );
-
-        if (!data) {
-          throw notFound(
-            `The service \`${service_slug}\` doesn't exist in this project.`
-          );
-        }
-        return data;
-      },
-      refetchInterval: (query) => {
-        if (query.state.data) {
-          return DEFAULT_QUERY_REFETCH_INTERVAL;
-        }
-        return false;
-      }
-    }),
-  deploymentList: ({
-    workspaceId,
-    project_slug,
-    service_slug,
-    env_slug,
-    filters = {}
-  }: {
-    workspaceId: string;
-    project_slug: string;
-    service_slug: string;
-    env_slug: string;
-    filters?: ServiceDeploymentListFilters;
-  }) =>
-    queryOptions({
-      queryKey: [
-        ...serviceQueries.single({
-          workspaceId,
-          project_slug,
-          service_slug,
-          env_slug
-        }).queryKey,
-        "DEPLOYMENT_LIST",
-        filters
-      ] as const,
-      queryFn: async ({ signal }) => {
-        const { data } = await apiClient.GET(
-          "/api/projects/{project_slug}/{env_slug}/service-details/{service_slug}/deployments/",
-          {
-            params: {
-              path: {
-                project_slug,
-                service_slug,
-                env_slug
-              },
-              query: {
-                ...filters,
-                queued_at_after: filters.queued_at_after?.toISOString(),
-                queued_at_before: filters.queued_at_before?.toISOString()
-              }
-            },
-            signal
-          }
-        );
-
-        if (!data) {
-          throw notFound();
-        }
-        return data;
-      },
-      refetchInterval: (query) => {
-        if (query.state.data) {
-          return DEFAULT_QUERY_REFETCH_INTERVAL;
-        }
-        return false;
-      }
-    }),
-  httpLogs: ({
-    workspaceId,
-    project_slug,
-    service_slug,
-    env_slug,
-    service_id,
-    autoRefetchEnabled = true,
-    filters = {},
-    queryClient
-  }: {
-    workspaceId: string;
-    service_id: string;
-    project_slug: string;
-    service_slug: string;
-    env_slug: string;
-    filters?: Omit<HTTPLogFilters, "isMaximized">;
-    queryClient: QueryClient;
-    autoRefetchEnabled?: boolean;
-  }) =>
-    infiniteQueryOptions({
-      queryKey: [
-        ...serviceQueries.single({
-          workspaceId,
-          project_slug,
-          service_slug,
-          env_slug
-        }).queryKey,
-        service_id,
-        "HTTP_LOGS",
-        filters
-      ] as const,
-      queryFn: async ({ pageParam, signal, queryKey }) => {
-        const allData = queryClient.getQueryData(queryKey) as InfiniteData<
-          HttpLogQueryData,
-          string | null
-        >;
-        const existingData = allData?.pages.find(
-          (_, index) => allData?.pageParams[index] === pageParam
-        );
-
-        /**
-         * We reuse the data in the query as we are sure this page is immutable,
-         * And we don't want to refetch the same logs that we have already fetched.
-         *
-         * However if we have the data in the cache and previous is `null`,
-         * it means that that page is the last and the next time we fetch it,
-         * it might have more data.
-         * Inspired by: https://github.com/TanStack/query/discussions/5921
-         */
-        if (existingData?.previous) {
-          return existingData;
-        }
-
-        let cursor = pageParam ?? undefined;
-        if (existingData?.cursor) {
-          cursor = existingData.cursor;
-        }
-
-        const { data } = await apiClient.GET("/api/http-logs/", {
-          params: {
-            query: {
-              ...filters,
-              cursor,
-              service_id,
-              per_page: DEFAULT_LOGS_PER_PAGE,
-              time_before: filters.time_before?.toISOString(),
-              time_after: filters.time_after?.toISOString()
-            }
-          },
-          signal
-        });
-
-        let apiData: HttpLogQueryData = {
-          next: null,
-          previous: null,
-          results: [],
-          cursor: null
-        };
-
-        if (data) {
-          const next = data.next
-            ? new URL(data.next).searchParams.get("cursor")
-            : null;
-          const previous = data.previous
-            ? new URL(data.previous).searchParams.get("cursor")
-            : null;
-          apiData = {
-            results: data.results,
-            next,
-            previous,
-            cursor: existingData?.cursor
-          };
-        }
-
-        // get cursor for initial page as its pageParam is `null`
-        // we want to do so that we don't to always fetch the latest data for the initial page
-        // instead what we want is to fetch from the data it starts
-        if (pageParam === null && apiData.next !== null && !apiData.cursor) {
-          const { data: nextPage } = await apiClient.GET("/api/http-logs/", {
-            params: {
-              query: {
-                ...filters,
-                service_id,
-                per_page: DEFAULT_LOGS_PER_PAGE,
-                cursor: apiData.next,
-                time_before: filters.time_before?.toISOString(),
-                time_after: filters.time_after?.toISOString()
-              }
-            },
-            signal
-          });
-          if (nextPage?.previous) {
-            apiData.cursor = new URL(nextPage.previous).searchParams.get(
-              "cursor"
-            );
-          }
-        }
-
-        return apiData;
-      },
-      refetchInterval: (query) => {
-        if (!query.state.data || !autoRefetchEnabled) {
-          return false;
-        }
-        return DEFAULT_QUERY_REFETCH_INTERVAL;
-      },
-      getNextPageParam: ({ next }) => next,
-      getPreviousPageParam: ({ previous }) => previous,
-      initialPageParam: null as string | null,
-      placeholderData: keepPreviousData,
-      staleTime: Number.POSITIVE_INFINITY
-    }),
-  metrics: ({
-    workspaceId,
-    project_slug,
-    service_slug,
-    env_slug,
-    filters
-  }: {
-    workspaceId: string;
-    project_slug: string;
-    service_slug: string;
-    env_slug: string;
-    filters?: MetricsFilters;
-  }) =>
-    queryOptions({
-      queryKey: [
-        ...serviceQueries.single({
-          workspaceId,
-          project_slug,
-          service_slug,
-          env_slug
-        }).queryKey,
-        "METRICS",
-        filters
-      ] as const,
-      queryFn: async ({ signal }) => {
-        const { data } = await apiClient.GET(
-          "/api/projects/{project_slug}/{env_slug}/service-details/{service_slug}/metrics/",
-          {
-            params: {
-              path: {
-                project_slug,
-                service_slug,
-                env_slug
-              },
-              query: {
-                ...filters
-              }
-            },
-            signal
-          }
-        );
-
-        if (!data) {
-          throw notFound();
-        }
-        return data;
-      },
-      refetchInterval: (query) => {
-        if (query.state.data) {
-          return DEFAULT_QUERY_REFETCH_INTERVAL;
-        }
-        return false;
-      }
-    }),
-  detectedPorts: ({
-    workspaceId,
-    project_slug,
-    service_slug,
-    env_slug
-  }: {
-    workspaceId: string;
-    project_slug: string;
-    service_slug: string;
-    env_slug: string;
-  }) =>
-    queryOptions({
-      queryKey: [
-        ...serviceQueries.single({
-          workspaceId,
-          project_slug,
-          service_slug,
-          env_slug
-        }).queryKey,
-        "DETECTED_PORTS"
-      ] as const,
-      queryFn: async ({ signal }) => {
-        const { data } = await apiClient.GET(
-          "/api/projects/{project_slug}/{env_slug}/service-details/{service_slug}/detected-ports/",
-          {
-            params: {
-              path: {
-                project_slug,
-                service_slug,
-                env_slug
-              }
-            },
-            signal
-          }
-        );
-
-        if (!data) {
-          throw notFound();
-        }
-        return data;
-      },
-      refetchInterval: (query) => {
-        if (query.state.data) {
-          return DEFAULT_QUERY_REFETCH_INTERVAL;
-        }
-        return false;
-      }
-    }),
-  availableVolumes: ({
-    workspaceId,
-    project_slug,
-    service_slug,
-    env_slug
-  }: {
-    workspaceId: string;
-    project_slug: string;
-    service_slug: string;
-    env_slug: string;
-  }) =>
-    queryOptions({
-      queryKey: [
-        ...serviceQueries.single({
-          workspaceId,
-          project_slug,
-          service_slug,
-          env_slug
-        }).queryKey,
-        "AVAILABLE_VOLUMES_FOR_SHARING"
-      ] as const,
-      queryFn: async ({ signal }) => {
-        const { data } = await apiClient.GET(
-          "/api/projects/{project_slug}/{env_slug}/{slug}/available-volumes/",
-          {
-            params: {
-              path: {
-                project_slug,
-                slug: service_slug,
-                env_slug
-              }
-            },
-            signal
-          }
-        );
-
-        if (!data) {
-          throw notFound();
-        }
-        return data;
-      },
-      placeholderData: keepPreviousData
-    }),
-  singleHttpLog: ({
-    workspaceId,
-    project_slug,
-    service_slug,
-    env_slug,
-    request_uuid
-  }: {
-    workspaceId: string;
-    project_slug: string;
-    service_slug: string;
-    env_slug: string;
-    service_id: string;
-    request_uuid: string;
-  }) =>
-    queryOptions({
-      queryKey: [
-        ...serviceQueries.single({
-          workspaceId,
-          project_slug,
-          service_slug,
-          env_slug
-        }).queryKey,
-        "HTTP_LOGS",
-        request_uuid
-      ] as const,
-      queryFn: async ({ signal }) => {
-        const { data } = await apiClient.GET("/api/http-logs/{request_uuid}/", {
-          params: {
-            path: {
-              request_uuid
-            }
-          },
-          signal
-        });
-        return data;
-      }
-    }),
-  filterHttpLogFields: ({
-    workspaceId,
-    project_slug,
-    service_slug,
-    env_slug,
-    service_id,
-    field,
-    value
-  }: {
-    workspaceId: string;
-    project_slug: string;
-    service_slug: string;
-    service_id: string;
-    env_slug: string;
-    field: RequestParams<"get", "/api/http-logs/fields/">["field"];
-    value: string;
-  }) =>
-    queryOptions({
-      queryKey: [
-        ...serviceQueries.single({
-          workspaceId,
-          project_slug,
-          service_slug,
-          env_slug
-        }).queryKey,
-        service_id,
-        "HTTP_LOG_FIELDS",
-        field,
-        value
-      ],
-      queryFn: async ({ signal }) => {
-        const { data } = await apiClient.GET("/api/http-logs/fields/", {
-          signal,
-          params: {
-            query: {
-              field,
-              value,
-              service_id
-            }
-          }
-        });
-        return data ?? [];
-      }
-    })
-};
-
-export const LOG_SOURCES = ["SYSTEM", "SERVICE"] as const;
-export const REQUEST_METHODS = [
-  "DELETE",
-  "GET",
-  "HEAD",
-  "OPTIONS",
-  "PATCH",
-  "POST",
-  "PUT"
-] as const;
-
-export const deploymentLogSearchSchema = zfd.formData({
-  level: zfd.repeatable(
-    z
-      .array(z.enum(LOG_LEVELS))
-      .optional()
-      .catch(LOG_LEVELS as Writeable<typeof LOG_LEVELS>)
-  ),
-  time_before: z.coerce.date().optional().catch(undefined),
-  time_after: z.coerce.date().optional().catch(undefined),
-  content: z.string().optional(),
-  query: z.string().optional(),
-  isMaximized: preprocess(
-    (arg) => arg === "true",
-    z.coerce.boolean().optional().catch(false)
-  ),
-  context: z.coerce.number().optional().catch(undefined),
-  context_lines: z.coerce.number().min(5).optional().catch(undefined)
-});
-
-export type DeploymentLogFilters = z.infer<typeof deploymentLogSearchSchema>;
-
-export const httpLogSearchSchema = zfd.formData({
-  time_before: z.coerce.date().optional().catch(undefined),
-  time_after: z.coerce.date().optional().catch(undefined),
-  request_method: zfd
-    .repeatable(z.array(z.enum(REQUEST_METHODS)).optional().catch(undefined))
-    .transform((val) => (val?.length === 0 ? undefined : val)),
-  request_query: z.string().optional(),
-  request_path: zfd
-    .repeatable(z.array(z.string()).optional().catch(undefined))
-    .transform((val) => (val?.length === 0 ? undefined : val)),
-  request_country_code: zfd
-    .repeatable(z.array(z.string()).optional().catch(undefined))
-    .transform((val) => (val?.length === 0 ? undefined : val)),
-  request_host: zfd
-    .repeatable(z.array(z.string()).optional().catch(undefined))
-    .transform((val) => (val?.length === 0 ? undefined : val)),
-  request_ip: zfd
-    .repeatable(z.array(z.string().ip()).optional().catch(undefined))
-    .transform((val) => (val?.length === 0 ? undefined : val)),
-  request_user_agent: zfd
-    .repeatable(z.array(z.string()).optional().catch(undefined))
-    .transform((val) => (val?.length === 0 ? undefined : val)),
-  request_id: z.string().uuid().optional().catch(undefined),
-  status: zfd
-    .repeatable(
-      z
-        .array(z.string())
-        .transform((array) =>
-          array.filter(
-            (val) =>
-              val.match(/\dxx/) || (!Number.isNaN(val) && Number(val) > 0)
-          )
-        )
-        .optional()
-        .catch(undefined)
-    )
-    .transform((val) => (val?.length === 0 ? undefined : val)),
-  isMaximized: preprocess(
-    (arg) => arg === "true",
-    z.coerce.boolean().optional().catch(false)
-  ),
-  sort_by: zfd
-    .repeatable(
-      z.array(
-        z.enum(["time", "-time", "request_duration_ns", "-request_duration_ns"])
-      )
-    )
-    .optional()
-    .catch(undefined)
-});
-
-export type HTTPLogFilters = z.infer<typeof httpLogSearchSchema>;
-
-export const deploymentQueries = {
-  recent: (workspaceId: string) =>
-    queryOptions({
-      queryKey: [...workspaceKey(workspaceId), "RECENT_DEPLOYMENTS"] as const,
-      queryFn: async ({ signal }) => {
-        const { data } = await apiClient.GET("/api/recent-deployments/", {
-          signal
-        });
-        if (!data) {
-          throw notFound(`This deployment does not exist in this service.`);
-        }
-        return data;
-      },
-      refetchInterval: (query) => {
-        if (query.state.data) {
-          return DEFAULT_QUERY_REFETCH_INTERVAL;
-        }
-        return false;
-      }
-    }),
-  single: ({
-    workspaceId,
-    project_slug,
-    service_slug,
-    env_slug,
-    deployment_hash
-  }: {
-    workspaceId: string;
-    project_slug: string;
-    service_slug: string;
-    env_slug: string;
-    deployment_hash: string;
-  }) =>
-    queryOptions({
-      queryKey: [
-        ...projectQueries.single(workspaceId, project_slug).queryKey,
-        env_slug,
-        "SERVICE_DETAILS",
-        service_slug,
-        "DEPLOYMENTS",
-        deployment_hash
-      ] as const,
-      queryFn: async ({ signal }) => {
-        const { data } = await apiClient.GET(
-          "/api/projects/{project_slug}/{env_slug}/service-details/{service_slug}/deployments/{deployment_hash}/",
-          {
-            params: {
-              path: {
-                project_slug,
-                service_slug,
-                env_slug,
-                deployment_hash
-              }
-            },
-            signal
-          }
-        );
-        if (!data) {
-          throw notFound(`This deployment does not exist in this service.`);
-        }
-        return data;
-      },
-      refetchInterval: (query) => {
-        if (query.state.data) {
-          return DEFAULT_QUERY_REFETCH_INTERVAL;
-        }
-        return false;
-      },
-      refetchIntervalInBackground: true
-    }),
-  logs: ({
-    workspaceId,
-    project_slug,
-    service_slug,
-    env_slug,
-    deployment_hash,
-    autoRefetchEnabled = true,
-    filters = {},
-    queryClient
-  }: {
-    workspaceId: string;
-    project_slug: string;
-    service_slug: string;
-    env_slug: string;
-    deployment_hash: string;
-    filters?: Omit<DeploymentLogFilters, "isMaximized">;
-    queryClient: QueryClient;
-    autoRefetchEnabled?: boolean;
-  }) =>
-    infiniteQueryOptions({
-      queryKey: [
-        ...deploymentQueries.single({
-          workspaceId,
-          project_slug,
-          service_slug,
-          env_slug,
-          deployment_hash
-        }).queryKey,
-        "RUNTIME_LOGS",
-        filters
-      ],
-      queryFn: async ({ pageParam, signal, queryKey }) => {
-        const allData = queryClient.getQueryData(queryKey) as InfiniteData<
-          DeploymentLogQueryData,
-          string | null
-        >;
-        const existingData = allData?.pages.find(
-          (_, index) => allData?.pageParams[index] === pageParam
-        );
-
-        /**
-         * We reuse the data in the query as we are sure this page is immutable,
-         * And we don't want to refetch the same logs that we have already fetched.
-         *
-         * However if we have the data in the cache and next is `null`,
-         * it means that that page is the last page with the most recent data
-         * and the next time we fetch it, there might be more data available.
-         * Inspired by: https://github.com/TanStack/query/discussions/5921
-         */
-        if (existingData?.next) {
-          return existingData;
-        }
-
-        /**
-         * when we issue a refetch, for all pages we fetched via `fetchPreviousPage` starting from the second page,
-         * tanstack query will use the `next` page pointer of the previous to refetch them,
-         * so we check if we already have it.
-         * In the docs, it's so that the data the pointers aren't stale, but we don't have that issue
-         * since the log data is immutable.
-         * ref: https://tanstack.com/query/latest/docs/framework/react/guides/infinite-queries#what-happens-when-an-infinite-query-needs-to-be-refetched
-         */
-        const existingDataIndex = allData?.pages.findIndex(
-          (_, index) => allData?.pages[index].next === pageParam
-        );
-        if (!existingData && existingDataIndex > -1) {
-          const nextPage = allData.pages[existingDataIndex + 1];
-          if (nextPage) {
-            return nextPage;
-          }
-        }
-
-        // the actual request
-        const { data } = await apiClient.GET(
-          "/api/projects/{project_slug}/{env_slug}/service-details/{service_slug}/deployments/{deployment_hash}/runtime-logs/",
-          {
-            params: {
-              path: {
-                project_slug,
-                service_slug,
-                env_slug,
-                deployment_hash
-              },
-              query: {
-                per_page: DEFAULT_LOGS_PER_PAGE,
-                cursor: pageParam ?? existingData?.cursor ?? undefined,
-                ...filters,
-                time_before: filters.time_before?.toISOString(),
-                time_after: filters.time_after?.toISOString()
-              }
-            },
-            signal
-          }
-        );
-
-        let apiData: DeploymentLogQueryData = {
-          next: null,
-          previous: null,
-          results: [],
-          cursor: null
-        };
-
-        if (data) {
-          // we reverse the results and reverse the page pointers (next/previous) because
-          // the data from the API is in reverse order of traversal and timestamp.
-          // Reversing them allows us to reorder the data in the ascending order as it is shown in the UI
-          apiData = {
-            results: data.results.toReversed(),
-            next: data?.previous ?? null,
-            previous: data?.next ?? null,
-            cursor: existingData?.cursor
-          };
-        }
-
-        // get cursor for initial page as its pageParam is `null`
-        // we want to do that because we don't to always fetch the latest data for the initial page
-        // instead what we want is to fetch from the time it starts
-        if (
-          pageParam === null &&
-          !apiData.cursor &&
-          !apiData.next &&
-          apiData.results.length > 0
-        ) {
-          const oldestLog = apiData.results[0];
-          const cursor = { sort: [oldestLog.timestamp], order: "asc" };
-          apiData.cursor = btoa(JSON.stringify(cursor));
-        }
-
-        return apiData;
-      },
-      getNextPageParam: ({ next }) => next,
-      getPreviousPageParam: ({ previous }) => previous,
-      initialPageParam: null as string | null,
-      refetchInterval: (query) => {
-        if (!query.state.data || !autoRefetchEnabled) {
-          return false;
-        }
-        return LOGS_QUERY_REFETCH_INTERVAL;
-      },
-      placeholderData: keepPreviousData,
-      staleTime: Number.POSITIVE_INFINITY
-    }),
-  logWithContext: ({
-    workspaceId,
-    project_slug,
-    service_slug,
-    env_slug,
-    deployment_hash,
-    time,
-    context_lines = 20
-  }: {
-    workspaceId: string;
-    project_slug: string;
-    service_slug: string;
-    env_slug: string;
-    deployment_hash: string;
-    time: number;
-    context_lines?: number;
-  }) =>
-    queryOptions({
-      queryKey: [
-        ...deploymentQueries.single({
-          workspaceId,
-          project_slug,
-          service_slug,
-          env_slug,
-          deployment_hash
-        }).queryKey,
-        "RUNTIME_LOGS",
-        "WITH_CONTEXT",
-        time,
-        context_lines
-      ],
-      queryFn: async ({ signal }) => {
-        const { data } = await apiClient.GET(
-          "/api/projects/{project_slug}/{env_slug}/service-details/{service_slug}/deployments/{deployment_hash}/runtime-logs/with-context/{time}",
-          {
-            params: {
-              path: {
-                project_slug,
-                service_slug,
-                env_slug,
-                deployment_hash,
-                time: time.toString()
-              },
-              query: {
-                lines: context_lines
-              }
-            },
-            signal
-          }
-        );
-
-        return data;
-      },
-      refetchInterval: (query) => {
-        if (!query.state.data) {
-          return false;
-        }
-        return LOGS_QUERY_REFETCH_INTERVAL;
-      },
-      placeholderData: keepPreviousData
-    }),
-  buildLogs: ({
-    workspaceId,
-    project_slug,
-    service_slug,
-    env_slug,
-    deployment_hash,
-    autoRefetchEnabled = true,
-    queryClient
-  }: {
-    workspaceId: string;
-    project_slug: string;
-    service_slug: string;
-    env_slug: string;
-    deployment_hash: string;
-    queryClient: QueryClient;
-    autoRefetchEnabled?: boolean;
-  }) =>
-    infiniteQueryOptions({
-      queryKey: [
-        ...deploymentQueries.single({
-          workspaceId,
-          project_slug,
-          service_slug,
-          env_slug,
-          deployment_hash
-        }).queryKey,
-        "BUILD_LOGS"
-      ],
-      queryFn: async ({ pageParam, signal, queryKey }) => {
-        const allData = queryClient.getQueryData(queryKey) as InfiniteData<
-          DeploymentLogQueryData,
-          string | null
-        >;
-        const existingData = allData?.pages.find(
-          (_, index) => allData?.pageParams[index] === pageParam
-        );
-
-        /**
-         * We reuse the data in the query as we are sure this page is immutable,
-         * And we don't want to refetch the same logs that we have already fetched.
-         *
-         * However if we have the data in the cache and next is `null`,
-         * it means that that page is the last page with the most recent data
-         * and the next time we fetch it, there might be more data available.
-         * Inspired by: https://github.com/TanStack/query/discussions/5921
-         */
-        if (existingData?.next) {
-          return existingData;
-        }
-
-        /**
-         * when we issue a refetch, for all pages we fetched via `fetchPreviousPage` starting from the second page,
-         * tanstack query will use the `next` page pointer of the previous to refetch them,
-         * so we check if we already have it.
-         * In the docs, it's so that the data the pointers aren't stale, but we don't have that issue
-         * since the log data is immutable.
-         * ref: https://tanstack.com/query/latest/docs/framework/react/guides/infinite-queries#what-happens-when-an-infinite-query-needs-to-be-refetched
-         */
-        const existingDataIndex = allData?.pages.findIndex(
-          (_, index) => allData?.pages[index].next === pageParam
-        );
-        if (!existingData && existingDataIndex > -1) {
-          const nextPage = allData.pages[existingDataIndex + 1];
-          if (nextPage) {
-            return nextPage;
-          }
-        }
-
-        // the actual request
-        const { data } = await apiClient.GET(
-          "/api/projects/{project_slug}/{env_slug}/service-details/{service_slug}/deployments/{deployment_hash}/build-logs/",
-          {
-            params: {
-              path: {
-                project_slug,
-                service_slug,
-                env_slug,
-                deployment_hash
-              },
-              query: {
-                per_page: DEFAULT_LOGS_PER_PAGE,
-                cursor: pageParam ?? existingData?.cursor ?? undefined
-              }
-            },
-            signal
-          }
-        );
-
-        let apiData: DeploymentLogQueryData = {
-          next: null,
-          previous: null,
-          results: [],
-          cursor: null
-        };
-
-        if (data) {
-          // we reverse the results and reverse the page pointers (next/previous) because
-          // the data from the API is in reverse order of traversal and timestamp.
-          // Reversing them allows us to reorder the data in the ascending order as it is shown in the UI
-          apiData = {
-            results: data.results.toReversed(),
-            next: data?.previous ?? null,
-            previous: data?.next ?? null,
-            cursor: existingData?.cursor
-          };
-        }
-
-        // get cursor for initial page as its pageParam is `null`
-        // we want to do that because we don't to always fetch the latest data for the initial page
-        // instead what we want is to fetch from the time it starts
-        if (
-          pageParam === null &&
-          !apiData.cursor &&
-          !apiData.next &&
-          apiData.results.length > 0
-        ) {
-          const oldestLog = apiData.results[0];
-          const cursor = { sort: [oldestLog.timestamp], order: "asc" };
-          apiData.cursor = btoa(JSON.stringify(cursor));
-        }
-
-        return apiData;
-      },
-      getNextPageParam: ({ next }) => next,
-      getPreviousPageParam: ({ previous }) => previous,
-      initialPageParam: null as string | null,
-      refetchInterval: (query) => {
-        if (!query.state.data || !autoRefetchEnabled) {
-          return false;
-        }
-        return LOGS_QUERY_REFETCH_INTERVAL;
-      },
-      placeholderData: keepPreviousData,
-      staleTime: Number.POSITIVE_INFINITY
-    }),
-  metrics: ({
-    workspaceId,
-    project_slug,
-    service_slug,
-    env_slug,
-    deployment_hash,
-    filters
-  }: {
-    workspaceId: string;
-    project_slug: string;
-    deployment_hash: string;
-    service_slug: string;
-    env_slug: string;
-    filters?: MetricsFilters;
-  }) =>
-    queryOptions({
-      queryKey: [
-        ...deploymentQueries.single({
-          workspaceId,
-          project_slug,
-          service_slug,
-          env_slug,
-          deployment_hash
-        }).queryKey,
-        "METRICS",
-        filters
-      ] as const,
-      queryFn: async ({ signal }) => {
-        const { data } = await apiClient.GET(
-          "/api/projects/{project_slug}/{env_slug}/service-details/{service_slug}/deployments/{deployment_hash}/metrics/",
-          {
-            params: {
-              path: {
-                project_slug,
-                service_slug,
-                env_slug,
-                deployment_hash
-              },
-              query: {
-                ...filters
-              }
-            },
-            signal
-          }
-        );
-
-        if (!data) {
-          throw notFound();
-        }
-        return data;
-      },
-      refetchInterval: (query) => {
-        if (query.state.data) {
-          return DEFAULT_QUERY_REFETCH_INTERVAL;
-        }
-        return false;
-      }
-    }),
-  httpLogs: ({
-    workspaceId,
-    project_slug,
-    service_slug,
-    env_slug,
-    deployment_hash,
-    autoRefetchEnabled = true,
-    filters = {},
-    queryClient
-  }: {
-    workspaceId: string;
-    project_slug: string;
-    service_slug: string;
-    env_slug: string;
-    deployment_hash: string;
-    filters?: Omit<HTTPLogFilters, "isMaximized">;
-    queryClient: QueryClient;
-    autoRefetchEnabled?: boolean;
-  }) =>
-    infiniteQueryOptions({
-      queryKey: [
-        ...deploymentQueries.single({
-          workspaceId,
-          project_slug,
-          service_slug,
-          env_slug,
-          deployment_hash
-        }).queryKey,
-        "HTTP_LOGS",
-        filters
-      ] as const,
-      queryFn: async ({ pageParam, signal, queryKey }) => {
-        const allData = queryClient.getQueryData(queryKey) as InfiniteData<
-          HttpLogQueryData,
-          string | null
-        >;
-        const existingData = allData?.pages.find(
-          (_, index) => allData?.pageParams[index] === pageParam
-        );
-
-        /**
-         * We reuse the data in the query as we are sure this page is immutable,
-         * And we don't want to refetch the same logs that we have already fetched.
-         *
-         * However if we have the data in the cache and previous is `null`,
-         * it means that that page is the last and the next time we fetch it,
-         * it might have more data.
-         * Inspired by: https://github.com/TanStack/query/discussions/5921
-         */
-        if (existingData?.previous) {
-          return existingData;
-        }
-
-        let cursor = pageParam ?? undefined;
-        if (existingData?.cursor) {
-          cursor = existingData.cursor;
-        }
-
-        const { data } = await apiClient.GET("/api/http-logs/", {
-          params: {
-            query: {
-              ...filters,
-              cursor,
-              deployment_id: deployment_hash,
-              per_page: DEFAULT_LOGS_PER_PAGE,
-              time_before: filters.time_before?.toISOString(),
-              time_after: filters.time_after?.toISOString()
-            }
-          },
-          signal
-        });
-
-        let apiData: HttpLogQueryData = {
-          next: null,
-          previous: null,
-          results: [],
-          cursor: null
-        };
-
-        if (data) {
-          const next = data.next
-            ? new URL(data.next).searchParams.get("cursor")
-            : null;
-          const previous = data.previous
-            ? new URL(data.previous).searchParams.get("cursor")
-            : null;
-          apiData = {
-            results: data.results,
-            next,
-            previous,
-            cursor: existingData?.cursor
-          };
-        }
-
-        // get cursor for initial page as its pageParam is `null`
-        // we want to do so that we don't to always fetch the latest data for the initial page
-        // instead what we want is to fetch from the data it starts
-        if (pageParam === null && apiData.next !== null && !apiData.cursor) {
-          const { data: nextPage } = await apiClient.GET("/api/http-logs/", {
-            params: {
-              query: {
-                ...filters,
-                per_page: DEFAULT_LOGS_PER_PAGE,
-                deployment_id: deployment_hash,
-                cursor: apiData.next,
-                time_before: filters.time_before?.toISOString(),
-                time_after: filters.time_after?.toISOString()
-              }
-            },
-            signal
-          });
-          if (nextPage?.previous) {
-            apiData.cursor = new URL(nextPage.previous).searchParams.get(
-              "cursor"
-            );
-          }
-        }
-
-        return apiData;
-      },
-      refetchInterval: (query) => {
-        if (!query.state.data || !autoRefetchEnabled) {
-          return false;
-        }
-        return DEFAULT_QUERY_REFETCH_INTERVAL;
-      },
-      getNextPageParam: ({ next }) => next,
-      getPreviousPageParam: ({ previous }) => previous,
-      initialPageParam: null as string | null,
-      placeholderData: keepPreviousData,
-      staleTime: Number.POSITIVE_INFINITY
-    }),
-  singleHttpLog: ({
-    workspaceId,
-    project_slug,
-    service_slug,
-    env_slug,
-    deployment_hash,
-    request_uuid
-  }: {
-    workspaceId: string;
-    project_slug: string;
-    service_slug: string;
-    env_slug: string;
-    deployment_hash: string;
-    request_uuid: string;
-  }) =>
-    queryOptions({
-      queryKey: [
-        ...deploymentQueries.single({
-          workspaceId,
-          project_slug,
-          service_slug,
-          env_slug,
-          deployment_hash
-        }).queryKey,
-        "HTTP_LOGS",
-        request_uuid
-      ] as const,
-      queryFn: async ({ signal }) => {
-        const { data } = await apiClient.GET("/api/http-logs/{request_uuid}/", {
-          params: {
-            path: {
-              request_uuid
-            }
-          },
-          signal
-        });
-        return data;
-      }
-    }),
-  filterHttpLogFields: ({
-    workspaceId,
-    project_slug,
-    service_slug,
-    env_slug,
-    deployment_hash,
-    field,
-    value
-  }: {
-    workspaceId: string;
-    project_slug: string;
-    service_slug: string;
-    env_slug: string;
-    deployment_hash: string;
-    field: RequestParams<"get", "/api/http-logs/fields/">["field"];
-    value: string;
-  }) =>
-    queryOptions({
-      queryKey: [
-        ...deploymentQueries.single({
-          workspaceId,
-          project_slug,
-          service_slug,
-          env_slug,
-          deployment_hash
-        }).queryKey,
-        "HTTP_LOG_FIELDS",
-        field,
-        value
-      ],
-      queryFn: async ({ signal }) => {
-        const { data } = await apiClient.GET("/api/http-logs/fields/", {
-          signal,
-          params: {
-            query: {
-              field,
-              value,
-              deployment_hash
-            }
-          }
-        });
-        return data ?? [];
-      }
-    })
-};
-
-export const serverQueries = {
-  settings: queryOptions({
-    queryKey: ["APP_SETTINGS"],
-    queryFn: async () => {
-      const { data } = await apiClient.GET("/api/settings/");
-      // we throw the error because we want ZaneOps to retry this for multiple times at least
-      return data;
-    },
-    staleTime: Number.MAX_SAFE_INTEGER
-  }),
-  resourceLimits: queryOptions({
-    queryKey: ["SERVICE_RESOURCE_LIMITS"],
-    queryFn: async () => {
-      const { data } = await apiClient.GET("/api/server/resource-limits/");
-      if (!data) throw new Error("Unknown error with the API");
-      return data;
-    },
-    staleTime: Number.MAX_SAFE_INTEGER
-  }),
-  ongoingUpdate: queryOptions({
-    queryKey: ["ONGOING_UPDATE"],
-    queryFn: async () => {
-      const { data } = await apiClient.GET("/api/check-ongoing-update-status/");
-      if (!data) throw new Error("Unknown error with the API");
-      return data;
-    }
-  })
-};
-
-type DeploymentLogQueryData = Pick<
-  NonNullable<
-    ApiResponse<
-      "get",
-      "/api/projects/{project_slug}/{env_slug}/service-details/{service_slug}/deployments/{deployment_hash}/runtime-logs/"
-    >
-  >,
-  "next" | "previous" | "results"
-> & {
-  cursor?: string | null;
-};
-
-export type HttpLogQueryData = Pick<
-  NonNullable<ApiResponse<"get", "/api/http-logs/">>,
-  "next" | "previous" | "results"
-> & {
-  cursor?: string | null;
-};
-
-export type HttpLogFilterField = RequestParams<
-  "get",
-  "/api/http-logs/fields/"
->["field"];
-
-export type DeploymentLog = Awaited<
-  ReturnType<
-    NonNullable<
-      Exclude<
-        ReturnType<typeof deploymentQueries.logs>["queryFn"],
-        typeof skipToken
-      >
-    >
-  >
->["results"][number];
-
-export type HttpLog = ApiResponse<"get", "/api/http-logs/{request_uuid}/">;
-
-export const resourceQueries = {
-  search: (workspaceId: string, query?: string) =>
-    queryOptions({
-      queryKey: [...workspaceKey(workspaceId), "RESOURCES", query] as const,
-      queryFn: async ({ signal }) => {
-        return apiClient.GET("/api/search-resources/", {
-          params: {
-            query: {
-              query: (query ?? "").trim()
-            }
-          },
-          signal
-        });
-      }
-    })
-};
-
-export type LatestRelease = {
-  tag: string;
-  url: string;
-  body: string;
-};
-
-export const versionQueries = {
-  latest: queryOptions<LatestRelease | null>({
-    queryKey: ["LATEST_RELEASE"] as const,
-    queryFn: async ({ signal }) => {
-      try {
-        const response = await fetch(
-          "https://cdn.zaneops.dev/api/latest-release",
-          { signal }
-        );
-        return response.json() as Promise<LatestRelease>;
-      } catch (error) {
-        return null;
-      }
-    },
-    refetchInterval: durationToMs(1, "hours")
-  })
-};
-
-export const sshKeysQueries = {
-  list: queryOptions({
-    queryKey: ["SSH_KEYS"] as const,
-    queryFn: async ({ signal }) => {
-      const { data } = await apiClient.GET("/api/shell/ssh-keys/", {
-        signal
-      });
-      if (!data) {
-        throw notFound("Oops !");
-      }
-      return data;
-    },
-    refetchInterval: (query) => {
-      if (query.state.data) {
-        return DEFAULT_QUERY_REFETCH_INTERVAL;
-      }
-      return false;
-    }
-  })
-};
-
-export const sharedRegistryCredentialsQueries = {
-  list: (workspaceId: string) =>
-    queryOptions({
-      queryKey: [
-        ...workspaceKey(workspaceId),
-        "SHARED_REGISTRY_CREDENTIALS"
-      ] as const,
-      queryFn: async ({ signal }) => {
-        const { data } = await apiClient.GET("/api/registries/credentials/", {
-          signal
-        });
-        if (!data) {
-          throw notFound("Oops !");
-        }
-        return data;
-      },
-      refetchInterval: (query) => {
-        if (query.state.data) {
-          return DEFAULT_QUERY_REFETCH_INTERVAL;
-        }
-        return false;
-      }
-    }),
-  single: (workspaceId: string, id: string) =>
-    queryOptions({
-      queryKey: [
-        ...workspaceKey(workspaceId),
-        "SHARED_REGISTRY_CREDENTIALS",
-        id
-      ] as const,
-      queryFn: async ({ signal }) => {
-        const { data } = await apiClient.GET(
-          "/api/registries/credentials/{id}/",
-          {
-            signal,
-            params: {
-              path: {
-                id
-              }
-            }
-          }
-        );
-        if (!data) {
-          throw notFound(
-            `No container registry credentials found with the ID ${id}`
-          );
-        }
-        return data;
-      },
-      refetchInterval: (query) => {
-        if (query.state.data) {
-          return DEFAULT_QUERY_REFETCH_INTERVAL;
-        }
-        return false;
-      }
-    })
-};
-
-export const gitAppsQueries = {
-  list: (workspaceId: string) =>
-    queryOptions({
-      queryKey: [...workspaceKey(workspaceId), "GIT_APPS"] as const,
-      queryFn: async ({ signal }) => {
-        const { data } = await apiClient.GET("/api/connectors/list/", {
-          signal
-        });
-        if (!data) {
-          throw notFound("Oops !");
-        }
-        return data;
-      },
-      refetchInterval: (query) => {
-        if (query.state.data) {
-          return DEFAULT_QUERY_REFETCH_INTERVAL;
-        }
-        return false;
-      }
-    }),
-  single: (workspaceId: string, id: string) =>
-    queryOptions({
-      queryKey: [...workspaceKey(workspaceId), "GIT_APPS", id] as const,
-      queryFn: async ({ signal }) => {
-        const { data } = await apiClient.GET("/api/connectors/{id}/", {
-          params: { path: { id } },
-          signal
-        });
-        if (!data) {
-          throw notFound("Oops !");
-        }
-        return data;
-      },
-      refetchInterval: (query) => {
-        if (query.state.data) {
-          return DEFAULT_QUERY_REFETCH_INTERVAL;
-        }
-        return false;
-      }
-    }),
-  github: (workspaceId: string, id: string) =>
-    queryOptions({
-      queryKey: [
-        ...gitAppsQueries.list(workspaceId).queryKey,
-        "GITHUB",
-        id
-      ] as const,
-      queryFn: async ({ signal }) => {
-        const { data } = await apiClient.GET("/api/connectors/github/{id}/", {
-          signal,
-          params: {
-            path: { id }
-          }
-        });
-        if (!data) {
-          throw notFound("This GitHub app does not exists.");
-        }
-        return data;
-      },
-      refetchInterval: (query) => {
-        if (query.state.data) {
-          return DEFAULT_QUERY_REFETCH_INTERVAL;
-        }
-        return false;
-      }
-    }),
-  gitlab: (workspaceId: string, id: string) =>
-    queryOptions({
-      queryKey: [
-        ...gitAppsQueries.list(workspaceId).queryKey,
-        "GITLAB",
-        id
-      ] as const,
-      queryFn: async ({ signal }) => {
-        const { data } = await apiClient.GET("/api/connectors/gitlab/{id}/", {
-          signal,
-          params: {
-            path: { id }
-          }
-        });
-        if (!data) {
-          throw notFound("This Gitlab app does not exists.");
-        }
-        return data;
-      },
-      refetchInterval: (query) => {
-        if (query.state.data) {
-          return DEFAULT_QUERY_REFETCH_INTERVAL;
-        }
-        return false;
-      }
-    }),
-  repositoryBranches: (
-    workspaceId: string,
-    repoUrl: string,
-    gitAppId?: string
-  ) =>
-    queryOptions({
-      queryKey: [
-        ...workspaceKey(workspaceId),
-        "GIT_REPOSITORY_BRANCHES",
-        repoUrl
-      ] as const,
-      queryFn: async ({ signal }) => {
-        const { data } = await apiClient.GET(
-          "/api/connectors/repository-branches/",
-          {
-            params: {
-              query: {
-                repository_url: repoUrl,
-                git_app_id: gitAppId
-              }
-            },
-            signal
-          }
-        );
-        return data ?? [];
-      },
-      enabled: () => z.string().url().safeParse(repoUrl).success, // only filter URLs to prevent errors when making requests in the backend
-      staleTime: durationToMs(30, "seconds"),
-      placeholderData: keepPreviousData
-    }),
-  repositories: (
-    workspaceId: string,
-    id: string,
-    filters: { query?: string } = {}
-  ) =>
-    queryOptions({
-      queryKey: [
-        ...gitAppsQueries.single(workspaceId, id).queryKey,
-        "REPOSITORIES",
-        filters
-      ] as const,
-      queryFn: async ({ signal }) => {
-        const { data } = await apiClient.GET(
-          "/api/connectors/{id}/repositories/",
-          {
-            params: {
-              path: {
-                id
-              },
-              query: {
-                // do not pass `filters.query` if empty
-                query: filters.query?.trim() ? filters.query.trim() : undefined
-              }
-            },
-            signal
-          }
-        );
-
-        if (!data) {
-          throw notFound("Oops !");
-        }
-
-        return data;
-      },
-      placeholderData: keepPreviousData
-    })
-};
-
-export const previewTemplatesQueries = {
-  list: (workspaceId: string, project_slug: string) =>
-    queryOptions({
-      queryKey: [
-        ...projectQueries.single(workspaceId, project_slug).queryKey,
-        "PREVIEW_TEMPLATES"
-      ] as const,
-      queryFn: async ({ signal }) => {
-        const { data } = await apiClient.GET(
-          "/api/projects/{slug}/preview-templates/",
-          {
-            signal,
-            params: {
-              path: {
-                slug: project_slug
-              }
-            }
-          }
-        );
-        if (!data) {
-          throw notFound("Oops !");
-        }
-        return data;
-      },
-      refetchInterval: (query) => {
-        if (query.state.data) {
-          return DEFAULT_QUERY_REFETCH_INTERVAL;
-        }
-        return false;
-      }
-    }),
-  single: (workspaceId: string, project_slug: string, template_slug: string) =>
-    queryOptions({
-      queryKey: [
-        ...previewTemplatesQueries.list(workspaceId, project_slug).queryKey,
-        template_slug
-      ] as const,
-      queryFn: async ({ signal }) => {
-        const { data } = await apiClient.GET(
-          "/api/projects/{project_slug}/preview-templates/{template_slug}/",
-          {
-            signal,
-            params: {
-              path: {
-                project_slug,
-                template_slug
-              }
-            }
-          }
-        );
-        if (!data) {
-          throw notFound("This preview template does not exist !");
-        }
-        return data;
-      },
-      refetchInterval: (query) => {
-        if (query.state.data) {
-          return DEFAULT_QUERY_REFETCH_INTERVAL;
-        }
-        return false;
-      }
-    })
-};
-
-export const buildRegistryListFilters = zfd.formData({
-  page: zfd.numeric().optional().catch(1).optional(),
-  per_page: zfd.numeric().optional().catch(10).optional()
-});
-
-export const buildRegistryImageListFilters = zfd.formData({
-  cursor: z.string().optional().catch(undefined)
-});
-
-export const buildRegistryQueries = {
-  list: (filters: z.infer<typeof buildRegistryListFilters>) =>
-    queryOptions({
-      queryKey: ["BUILD_REGISTRY_CREDENTIALS", filters] as const,
-      queryFn: async ({ signal }) => {
-        const { data } = await apiClient.GET(
-          "/api/registries/build-registries/",
-          {
-            signal,
-            params: {
-              query: filters
-            }
-          }
-        );
-        if (!data) {
-          throw notFound("Oops !");
-        }
-        return data;
-      },
-      refetchInterval: (query) => {
-        if (query.state.data) {
-          return DEFAULT_QUERY_REFETCH_INTERVAL;
-        }
-        return false;
-      }
-    }),
-  single: (id: string) =>
-    queryOptions({
-      queryKey: ["BUILD_REGISTRY_CREDENTIALS", id] as const,
-      queryFn: async ({ signal }) => {
-        const { data } = await apiClient.GET(
-          "/api/registries/build-registries/{id}/",
-          {
-            signal,
-            params: {
-              path: {
-                id
-              }
-            }
-          }
-        );
-        if (!data) {
-          throw notFound(
-            `No build registry credentials found with the ID ${id}`
-          );
-        }
-        return data;
-      },
-      refetchInterval: (query) => {
-        if (query.state.data) {
-          return DEFAULT_QUERY_REFETCH_INTERVAL;
-        }
-        return false;
-      }
-    }),
-  imageList: (
-    id: string,
-    params: z.infer<typeof buildRegistryImageListFilters>
-  ) =>
-    queryOptions({
-      queryKey: [
-        ...buildRegistryQueries.single(id).queryKey,
-        "IMAGE_LIST",
-        params
-      ] as const,
-      queryFn: async ({ signal }) => {
-        const { data } = await apiClient.GET(
-          "/api/registries/build-registries/{id}/list-images/",
-          {
-            params: {
-              path: {
-                id
-              },
-              query: params
-            },
-            signal
-          }
-        );
-
-        if (!data) throw new Error("Not found");
-        return data;
-      },
-      refetchInterval: (query) => {
-        if (!query.state.data) {
-          return false;
-        }
-        return DEFAULT_QUERY_REFETCH_INTERVAL;
-      },
-      placeholderData: keepPreviousData
-    })
-};
-
-/************************************
- *       Stack Template Queries     *
- ************************************/
-
-export const templateSearchFilters = zfd.formData({
-  query: z.string().optional().catch(""),
-  page: zfd.numeric().catch(1).optional().default(1),
-  perPage: zfd.numeric().optional().catch(15).optional().default(15),
-  tags: zfd.repeatable(z.array(z.string())).optional().default([])
-});
-
-export type TemplateSearchFilters = z.infer<typeof templateSearchFilters>;
-
-export const templateQueries = {
-  search: (filters: TemplateSearchFilters) =>
-    queryOptions({
-      queryKey: ["TEMPLATE_SEARCH", filters],
-      queryFn: async ({ signal }) => {
-        const url = new URL("/api/search", TEMPLATE_API_HOST);
-
-        if (filters.query) {
-          url.searchParams.set("q", filters.query);
-        }
-
-        if (filters.page) {
-          url.searchParams.set("page", filters.page.toString());
-        }
-
-        if (filters.perPage) {
-          url.searchParams.set("per_page", filters.perPage.toString());
-        }
-
-        for (const tag of filters.tags) {
-          url.searchParams.append("tags", tag);
-        }
-
-        const response = await fetch(url, { signal });
-
-        if (!response.ok) {
-          throw new Error("Failed to fetch templates");
-        }
-
-        return response.json() as Promise<TemplateSearchAPIResponse>;
-      }
-    }),
-  tags: queryOptions({
-    queryKey: ["TEMPLATE_SEARCH", "TAGS"],
-    queryFn: async ({ signal }) => {
-      const url = new URL("/api/tags.json", TEMPLATE_API_HOST);
-      const response = await fetch(url, { signal });
-
-      if (!response.ok) {
-        throw new Error("Failed to fetch tags");
-      }
-
-      return response.json() as Promise<string[]>;
-    }
-  }),
-  single: (templateSlug: string) =>
-    queryOptions({
-      queryKey: ["TEMPLATE_SEARCH", templateSlug],
-      queryFn: async ({ signal }) => {
-        const url = new URL(
-          `/api/templates/${templateSlug}.json`,
-          TEMPLATE_API_HOST
-        );
-        const response = await fetch(url, { signal });
-
-        if (!response.ok) {
-          throw notFound("This template doesn't exist on ZaneOps");
-        }
-
-        return response.json() as Promise<TemplateDetailsApiResponse>;
-      }
-    })
-};
-
-/************************************
- *         Workspace Queries        *
- ************************************/
-
-export const workspaceMemberListFilters = zfd.formData({
-  page: zfd.numeric().optional().catch(1).optional(),
-  query: z.string().optional(),
-  per_page: zfd.numeric().optional().catch(10).optional(),
-  role: z
-    .enum(["Viewer", "Member", "Admin", "Owner"])
-    .optional()
-    .catch(undefined)
-});
-
-export const paginationListFilters = zfd.formData({
-  page: zfd.numeric().optional().catch(1).optional(),
-  per_page: zfd.numeric().optional().catch(10).optional()
-});
-
-export const workspaceQueries = {
-  members: (
-    workspaceId: string,
-    filters: z.infer<typeof workspaceMemberListFilters> = {}
-  ) =>
-    queryOptions({
-      queryKey: [...workspaceKey(workspaceId), "MEMBERS", filters] as const,
-      queryFn: async ({ signal }) => {
-        const { data } = await apiClient.GET("/api/workspace/members/", {
-          signal,
-          params: {
-            query: {
-              ...filters,
-              role: filters.role
-                ? WORKSPACE_ROLE_MAPPING[filters.role].value
-                : undefined
-            }
-          }
-        });
-        if (!data) throw notFound("Not found");
-        return data;
-      },
-      refetchInterval: (query) => {
-        if (!query.state.data) {
-          return false;
-        }
-        return DEFAULT_QUERY_REFETCH_INTERVAL;
-      },
-      placeholderData: keepPreviousData
-    }),
-  invitations: (
-    workspaceId: string,
-    filters: z.infer<typeof paginationListFilters> = {}
-  ) =>
-    queryOptions({
-      queryKey: [...workspaceKey(workspaceId), "INVITATIONS", filters] as const,
-      queryFn: async ({ signal }) => {
-        const { data } = await apiClient.GET("/api/workspace/invitations/", {
-          signal,
-          params: {
-            query: filters
-          }
-        });
-        if (!data) throw notFound("Not found");
-        return data;
-      },
-      refetchInterval: (query) => {
-        if (!query.state.data) {
-          return false;
-        }
-        return DEFAULT_QUERY_REFETCH_INTERVAL;
-      },
-      placeholderData: keepPreviousData
-    }),
-
-  invitationLink: (token: string) =>
-    queryOptions({
-      queryKey: ["INVITATION_LINKS", token],
-      queryFn: async ({ signal }) => {
-        const result = await apiClient.GET(
-          "/api/workspace/invitations/{token}/",
-          {
-            signal,
-            params: {
-              path: { token }
-            }
-          }
-        );
-
-        // if rate limited, throw error
-        if (result.response.status === 429) {
-          const fullErrorMessage = result.error?.errors
-            .map((err) => err.detail)
-            .join(" ");
-
-          throw new Error(fullErrorMessage);
-        }
-
-        if (!result.data) {
-          const fullErrorMessage = result.error?.errors
-            .map((err) => err.detail)
-            .join(" ");
-
-          throw notFound(fullErrorMessage);
-        }
-
-        return result.data;
-      }
-    }),
-
-  member: (workspaceId: string, membershipId: string) =>
-    queryOptions({
-      queryKey: [
-        ...workspaceKey(workspaceId),
-        "MEMBERS",
-        "SINGLE",
-        membershipId
-      ] as const,
-      queryFn: async ({ signal }) => {
-        const { data } = await apiClient.GET(
-          "/api/workspace/members/{membership_id}/",
-          {
-            signal,
-            params: {
-              path: {
-                membership_id: membershipId
-              }
-            }
-          }
-        );
-        if (!data) throw notFound("Not found");
-        return data;
-      },
-      refetchInterval: (query) => {
-        if (!query.state.data) {
-          return false;
-        }
-        return DEFAULT_QUERY_REFETCH_INTERVAL;
-      },
-      placeholderData: keepPreviousData
-    })
-};
